fix datatype

+++ database/migrations/2022_06_15_162221_add_second_hour_rate_column_to_zones_table.php
@@ -14,7 +14,7 @@ class AddSecondHourRateColumnToZonesTable extends Migration
     public function up()
     {
         Schema::table('zones', function (Blueprint $table) {
-            $table->integer('second_hour_rate')->nullable()->after('hour_rate');
+            $table->float('second_hour_rate')->nullable()->after('hour_rate');
         });
     }
